requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
instagrapi
python-dotenv
google-genai
google-generativeai
anthropic
playwright
requests
openai
discord.py
flask
flask-cors
